Component entity filtering (#28)
Closes #23.

Changes the signature of defineComponent to take a dictionary of arguments instead of a tuple. Also switches to using t to validate this. In the future, t will be used in more places. The first argument, the component class name, is specified with the name field; the generator function is specified with the generator field. Both are mandatory.

Using the new dictionary table, there is a new optional argument, entityFilter. It takes an entity (whether abstract entity ID or Roblox instance) and should return a bool. If it returns false for a given entity, addComponent and batchAddComponents will throw when trying to add the component to that entity. Otherwise, the addition proceeds as normal. batchAddComponents throws if any component cannot be added to the entity.